First Commit
I have utilized the Netflix dataset, available in CSV format, to create an interactive Dashboard that provides various insights. The insights include:

1. The top 20 genres represented in the dataset.
2. The total count of movies and TV shows released by each country.
3. The distribution of movies and TV shows released by year.
4. The most commonly assigned rating for movies and TV shows.

To visualize these insights, I have designed a Tableau Dashboard.
You can interact with the dashboard to explore the data and find the movie information you're looking for: